0519_오전

+++ src/semi/comments/Comments_deleteController.java
@@ -15,15 +15,21 @@ import org.json.JSONObject;
 public class Comments_deleteController extends HttpServlet {
 	@Override
 	protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		req.setCharacterEncoding("utf-8");
+		resp.setCharacterEncoding("text/plain;charset=utf-8");
 		String scomments_num=req.getParameter("comments_num");
 		int comments_num=Integer.parseInt(scomments_num);
 		CommentsDao dao=CommentsDao.getInstance();
-		int n=dao.delete_comments(comments_num);
+		String susers_id=req.getParameter("users_id");
+		String my_id=(String)req.getSession().getAttribute("users_id");
 		boolean result=false;
-		if(n>0) {
-			result=true;
+		if(susers_id.equals(my_id)) {
+			int n=dao.delete_comments(comments_num);
+			if(n>0) {
+				result=true;
+			}
 		}
-		resp.setCharacterEncoding("text/plain;charset=utf-8");
+		
 		JSONObject json=new JSONObject();
 		json.put("result", result);
 		PrintWriter pw=resp.getWriter();
